requirements.txt
jupyterlab>3.0
ninja==1.11.1.1
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
torch-optimi==0.2.1
zstd==1.5.5.1
datasets==2.21.0
evaluate==0.4.1
huggingface_hub==0.24.6
tokenizers==0.19.1
transformers==4.44.1
safetensors==0.4.4
wandb==0.16.6
